requirements.txt
Django==4.1.1
django-environ==0.9.0
django-debug-toolbar==3.6.0
